amended landing page
